% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency.R
\name{adjacency}
\alias{adjacency}
\title{Adjaceny matrix}
\usage{
adjacency(df)
}
\arguments{
\item{name}{String of the column used for subsetting.}

\item{pattern}{String to match.}
}
\value{
a data.frame with network metrics.
}
\description{
Adjaceny matrix
}
