% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-foodwebs.R
\name{api_foodwebs}
\alias{api_foodwebs}
\title{Create API URL for Food Webs}
\usage{
api_foodwebs(
  ecosystemType = NULL,
  xmin = NULL,
  xmax = NULL,
  ymin = NULL,
  ymax = NULL
)
}
\arguments{
\item{ecosystemType}{Character of ecosystem type.}

\item{xmin}{Numeric value of minimum longitude.}

\item{xmax}{Numeric value of maximum longitude.}

\item{ymin}{Numeric value of minimum latitude.}

\item{ymax}{Numeric value of maximum latitude.}
}
\value{
Data frame with food web data.
}
\description{
Create API URL for Food Webs
}
\details{
Arguments are used to filter the food webs. When
no parameters are provided, all food webs are returned.
}
