% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-taxa.R
\name{api_taxa}
\alias{api_taxa}
\title{Create API URL for Taxa}
\usage{
api_taxa(taxonRank = NULL, taxonomicStatus = NULL)
}
\arguments{
\item{taxonRank}{Character.}

\item{taxonomicStatus}{Character.}
}
\value{
Data frame with taxon data.
}
\description{
Create API URL for Taxa
}
\details{
Arguments are used to filter the taxa. When
no parameters are provided, all taxa are returned.
}
