% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-communities.R
\name{.api_communities}
\alias{.api_communities}
\title{Create API URL for Communties}
\usage{
.api_communities(params = NULL)
}
\arguments{
\item{params}{list of filtering parameters.}
}
\value{
Data frame with food web data.
}
\description{
Create API URL for Communties
}
\details{
Arguments are used to filter the communities. When
no parameters are provided, all communities are returned.
}
