% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-foodwebs.R
\name{.api_foodwebs}
\alias{.api_foodwebs}
\title{Create API URL for Food Webs}
\usage{
.api_foodwebs(params = NULL)
}
\arguments{
\item{params}{list of filtering parameters.}
}
\value{
Data frame with food web data.
}
\description{
Create API URL for Food Webs
}
\details{
Arguments are used to filter the food webs. When
no parameters are provided, all food webs are returned.
}
