% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-interactions.R
\name{.api_interactions}
\alias{.api_interactions}
\title{Create API URL for Interactions}
\usage{
.api_interactions(params = NULL)
}
\arguments{
\item{params}{list of filtering parameters.}
}
\value{
Data frame with food web data.
}
\description{
Create API URL for Interactions
}
\details{
Arguments are used to filter the food webs. When
no parameters are provided, all food webs are returned.
}
