% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-communities.R
\name{.join_communities}
\alias{.join_communities}
\title{Join Raw Community Table with other Tables}
\usage{
.join_communities(df)
}
\arguments{
\item{df}{Table with raw community data.}
}
\value{
Data frame with community data.
}
\description{
Join Raw Community Table with other Tables
}
