% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-interactions.R
\name{.join_interactions}
\alias{.join_interactions}
\title{Join Raw Interaction Table with other Tables}
\usage{
.join_interactions(df)
}
\arguments{
\item{df}{Table with raw community data.}
}
\value{
Data frame with community data.
}
\description{
Join Raw Interaction Table with other Tables
}
