% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gateway}
\alias{gateway}
\title{GATEWAy Food Web Database}
\format{
\subsection{\code{gateway}}{

A data frame with 222,151 rows and 46 columns.
}
}
\source{
\url{https://doi.org/10.25829/idiv.283-3-756}
}
\usage{
gateway
}
\description{
The GATEWAy database from \url{https://doi.org/10.25829/idiv.283-3-756}
}
\keyword{datasets}
