% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-communities.R
\name{gateway_communities}
\alias{gateway_communities}
\title{Download Food Webs}
\usage{
gateway_communities(
  foodwebID = NULL,
  columns = c("foodwebName", "acceptedTaxonName", "lifeStage", "meanMass", "meanLength",
    "biomass")
)
}
\arguments{
\item{foodwebID}{Numeric value of Food Web ID.}

\item{columns}{Character vector of the columns to return.}
}
\value{
Data frame with food web data.
}
\description{
Download Food Webs
}
\details{
Arguments are used to filter the communities. When
no parameters are provided, all communities are returned..
}
