% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-interactions.R
\name{gateway_interactions}
\alias{gateway_interactions}
\title{Download Interactions}
\usage{
gateway_interactions(foodwebID = NULL, resourceID = NULL, consumerID = NULL)
}
\arguments{
\item{foodwebID}{Integer of Food Web ID.}

\item{resourceID}{Integer of Resource ID.}

\item{consumerID}{Integer of Consumer ID.}
}
\value{
Data frame with interaction data.
}
\description{
Download Interactions
}
\details{
Arguments are used to filter the interactions. When
no parameters are provided, all interactions are returned.
}
