% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-taxa.R
\name{gateway_taxa}
\alias{gateway_taxa}
\title{Download Taxa}
\usage{
gateway_taxa(taxonRank = NULL, taxonomicStatus = NULL)
}
\value{
Data frame with food web data.
}
\description{
Download Taxa
}
\details{
Arguments are used to filter the food webs. When
no parameters are provided, all food webs are returned.
}
