% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-communities.R
\name{get_community}
\alias{get_community}
\title{Download GATEWAy Communities}
\usage{
get_community(
  foodwebID = NULL,
  columns = c("foodwebName", "acceptedTaxonName", "lifeStage", "meanMass", "meanLength",
    "biomass")
)
}
\arguments{
\item{foodwebID}{Numeric value of Food Web ID.}

\item{columns}{Character vector of the columns to return.}
}
\value{
Data frame with food web data.
}
\description{
Download GATEWAy Communities
}
\details{
Arguments are used to filter the communities. When
no parameters are provided, all communities are returned..
}
