% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-foodwebs.R
\name{get_foodweb}
\alias{get_foodweb}
\title{Download Food Webs}
\usage{
get_foodweb(
  foodwebID = NULL,
  ecosystemType = NULL,
  xmin = NULL,
  xmax = NULL,
  ymin = NULL,
  ymax = NULL
)
}
\arguments{
\item{foodwebID}{Integer of Food Web ID.}

\item{ecosystemType}{Character of ecosystem type.}

\item{xmin}{Numeric value of minimum longitude.}

\item{xmax}{Numeric value of maximum longitude.}

\item{ymin}{Numeric value of minimum latitude.}

\item{ymax}{Numeric value of maximum latitude.}
}
\value{
Data frame with food web data.
}
\description{
Download Food Webs
}
\details{
Arguments are used to filter the food webs. When
no parameters are provided, all food webs are returned.
}
