% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-interaction-methods.R
\name{get_interaction_method}
\alias{get_interaction_method}
\title{GATEWAy Interaction Methods}
\usage{
get_interaction_method()
}
\value{
Character vector.
}
\description{
GATEWAy Interaction Methods
}
