% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-taxa.R
\name{get_taxon}
\alias{get_taxon}
\title{Download Taxa}
\usage{
get_taxon(taxonID = NULL, taxonRank = NULL, taxonomicStatus = NULL)
}
\arguments{
\item{taxonID}{Integer of Taxon ID.}

\item{taxonRank}{Character.}

\item{taxonomicStatus}{Character.}
}
\value{
Data frame with food web data.
}
\description{
Download Taxa
}
\details{
Arguments are used to filter the food webs. When
no parameters are provided, all food webs are returned.
}
