% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network-dissimilarity.R
\name{network_dissimilarity}
\alias{network_dissimilarity}
\title{Network dissimlarities}
\usage{
network_dissimilarity(df)
}
\arguments{
\item{df}{data.frame subset from GATEWAy.}
}
\value{
Named matrix with the network disimilarity between food webs.
}
\description{
Network dissimlarities
}
