% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network-metrics.R
\name{network_metrics}
\alias{network_metrics}
\title{Compute network metrics}
\usage{
network_metrics(fw)
}
\arguments{
\item{fw}{matrix, adjacency matrix of the food web with resources}
}
\value{
a data.frame with network metrics.
}
\description{
Compute network metrics
}
