% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network-dissimilarity.R
\name{network_pca}
\alias{network_pca}
\title{Network PCA}
\usage{
network_pca(df, ...)
}
\arguments{
\item{df}{data.frame subset from GATEWAy.}

\item{...}{additional parameters for prcomp.}
}
\value{
Named matrix with the network disimilarity between food webs.
}
\description{
Network PCA
}
