% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccard.R
\name{species_jaccard}
\alias{species_jaccard}
\title{Compute species (Jaccard) similarity}
\usage{
species_jaccard(df)
}
\arguments{
\item{df}{data.frame subset from GATEWAy.}
}
\value{
Named matrix with the Jaccard similarity between food webs.
}
\description{
Compute species (Jaccard) similarity
}
