% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-foodwebs.R
\name{gateway_foodwebs}
\alias{gateway_foodwebs}
\title{Download All Food Webs}
\usage{
gateway_foodwebs(
  ecosystemType = NULL,
  xmin = NULL,
  xmax = NULL,
  ymin = NULL,
  ymax = NULL
)
}
\arguments{
\item{ecosystemType}{Character of ecosystem type.}

\item{xmin}{Nnumeric value of minimum longitude.}

\item{xmax}{Nnumeric value of maximum longitude.}

\item{ymin}{Nnumeric value of minimum latitude.}

\item{ymax}{Nnumeric value of maximum latitude.}
}
\value{
Data frame with food web data.
}
\description{
Download All Food Webs
}
\details{
Arguments are used to filter the food webs. When
no parameters are provided, all food webs are returned.
}
